+++ py_trees_ros/visitors.py
@@ -27,12 +27,48 @@ runs its own method on the behaviour to do as it wishes - logging, introspecting
 
 import py_trees.visitors
 import py_trees_ros_interfaces.msg as py_trees_msgs
+import rclpy
+import time
 
 from . import conversions
 
 ##############################################################################
 # Visitors
 ##############################################################################
+
+
+class SetupLogger(py_trees.visitors.VisitorBase):
+    """
+    Use as a visitor to :meth:`py_trees_ros.trees.TreeManager.setup`
+    to log the name and timings of each behaviours' setup
+    to the ROS debug channel.
+
+    Args:
+        node: an rclpy node that will provide debug logger
+    """
+    def __init__(self, node: rclpy.node.Node):
+        super().__init__(full=True)
+        self.node = node
+
+    def initialise(self):
+        """
+        Initialise the timestamping chain.
+        """
+        self.start_time = time.monotonic()
+        self.last_time = self.start_time
+
+    def run(self, behaviour):
+        current_time = time.monotonic()
+        self.node.get_logger().debug(
+            "'{}'.setup: {:.4f}s".format(behaviour.name, current_time - self.last_time)
+        )
+        self.last_time = current_time
+
+    def finalise(self):
+        current_time = time.monotonic()
+        self.node.get_logger().debug(
+            "Total tree setup time: {:.4f}s".format(current_time - self.start_time)
+        )
 
 
 class TreeToMsgVisitor(py_trees.visitors.VisitorBase):